List available replications for a paper

Retrieves the replication metadata for a paper and lists all available figures and tables that can be reproduced.

Usage

list_replications(doi)

Arguments

doi

Character. DOI of the paper.

Value

A list of replication identifiers.

Examples

if (FALSE) { # \dontrun{
list_replications("10.1177/00491241211036161")
} # }