mermaid.html

replicateEverything: Reproduce Empirical Research Results

The replicateEverything package provides tools for discovering and executing computational replications of empirical research papers. It connects to a public replication registry containing metadata, replication scripts, and processed datasets required to reproduce figures and tables from published studies.

Workflow

A typical workflow using the package is:

Retrieve metadata for a paper using get_doi_metadata().

Search the replication registry using search_papers().

Create a template for contributing a replication using create_replication_template().

Inspect available replications using list_replications().

Run a single replication using run_replication().

Reproduce all results from a paper using replicate_paper().

Examples

Retrieve metadata for a paper:

get_doi_metadata("10.1177/00491241211036161")

Search the registry:

search_papers("causal")

Create a replication template:

create_replication_template("10.1177/00491241211036161")

List replications:

list_replications("10.1177/00491241211036161")

Run a single replication:

run_replication("10.1177/00491241211036161","fig_1")

Replicate an entire paper:

replicate_paper("10.1177/00491241211036161")

Registry

Replication metadata and materials are stored in the public registry: https://github.com/replicate-anything/registry.

See also

Author

Maintainer: Vermon Washington vermon.washington@wzb.eu

Authors:

Macartan Humphreys macartan.humphreys@wzb.eu